requirements-dev.txt
# Testing tools
pytest==3.2.3 # rq.filter: <4.0
pyIATI==0.2.0 # rq.filter: ==0.2.0
